Switching doubles to floats in inputMonitor
Changes:
* Input controller doesn't require the accuracy of doubles, so I switched them for floats. This could help embedded performance a bit.

+++ src/main/engine/Misc/src/inputMonitor.cpp
@@ -14,7 +14,7 @@
 const int JOYSTICK_DEAD_ZONE = 4000;
 #define PI 3.14159265
 /// @todo - Use event based input handling; DO NOT REFACTOR THIS FILE
-vector<double> cameraDistance(vec3 offset);
+vector<float> cameraDistance(vec3 offset);
 
 /*
  (void) rotateShape takes a (void *) gameInfoStruct that should be of type
@@ -95,11 +95,11 @@ void rotateShape(void *gameInfoStruct, void *target) {
             if (mouseY < 0) {
                 modifier = (mouseY / 5.0f) * -1;
             } else if (controllerRightStateY < -JOYSTICK_DEAD_ZONE) {
-                modifier = (static_cast<double>(controllerRightStateY * -1)) / INT16_MAX;
+                modifier = (static_cast<float>(controllerRightStateY * -1)) / INT16_MAX;
             }
-            vector<double> distHold = cameraDistance(cameraOffset);
+            vector<float> distHold = cameraDistance(cameraOffset);
             cameraOffset[1] -= offsetSpeed * modifier;
-            vector<double> distFinish = cameraDistance(cameraOffset);
+            vector<float> distFinish = cameraDistance(cameraOffset);
             distHold[0] = sqrt(distHold[0]);
             distHold[1] = sqrt(distHold[1]);
             distFinish[0] = sqrt(distFinish[0]);
@@ -119,9 +119,9 @@ void rotateShape(void *gameInfoStruct, void *target) {
             } else if (controllerRightStateY > JOYSTICK_DEAD_ZONE) {
                 modifier = static_cast<float>(controllerRightStateY) / INT16_MAX;
             }
-            vector<double> distHold = cameraDistance(cameraOffset);
+            vector<float> distHold = cameraDistance(cameraOffset);
             cameraOffset[1] += offsetSpeed * modifier;
-            vector<double> distFinish = cameraDistance(cameraOffset);
+            vector<float> distFinish = cameraDistance(cameraOffset);
             distHold[0] = sqrt(distHold[0]);
             distHold[1] = sqrt(distHold[1]);
             distFinish[0] = sqrt(distFinish[0]);
@@ -141,7 +141,7 @@ void rotateShape(void *gameInfoStruct, void *target) {
         if (currentGame->getKeystate()[SDL_SCANCODE_KP_7] ||
             (mouseX < 0 && trackMouse) || controllerRightStateX < -JOYSTICK_DEAD_ZONE) {
             // Rotate the camera about the y axis
-            double distHold = cameraOffset[0] * cameraOffset[0] + cameraOffset[2] * cameraOffset[2];
+            float distHold = cameraOffset[0] * cameraOffset[0] + cameraOffset[2] * cameraOffset[2];
             float multiplier = 1.0f;
             if (mouseX < 0) {
                 multiplier = (mouseX * -1.0f) / 5.0f;
@@ -161,7 +161,7 @@ void rotateShape(void *gameInfoStruct, void *target) {
                 cameraOffset[0] -= offsetSpeed * multiplier;
                 cameraOffset[2] += offsetSpeed * multiplier;
             }
-            double distFinish = cameraOffset[0] * cameraOffset[0] + cameraOffset[2] * cameraOffset[2];
+            float distFinish = cameraOffset[0] * cameraOffset[0] + cameraOffset[2] * cameraOffset[2];
             distHold = sqrt(distHold);
             distFinish = sqrt(distFinish);
             distFinish /= distHold;
@@ -170,7 +170,7 @@ void rotateShape(void *gameInfoStruct, void *target) {
         }
         if (currentGame->getKeystate()[SDL_SCANCODE_KP_9] ||
             (mouseX > 0 && trackMouse) || controllerRightStateX > JOYSTICK_DEAD_ZONE) {
-            double distHold = cameraOffset[0] * cameraOffset[0] + cameraOffset[2] * cameraOffset[2];
+            float distHold = cameraOffset[0] * cameraOffset[0] + cameraOffset[2] * cameraOffset[2];
             float multiplier = 1.0f;
             if (mouseX > 0) {
                 multiplier = mouseX / 5.0f;
@@ -190,7 +190,7 @@ void rotateShape(void *gameInfoStruct, void *target) {
                 cameraOffset[0] += offsetSpeed * multiplier;
                 cameraOffset[2] -= offsetSpeed * multiplier;
             }
-            double distFinish = cameraOffset[0] * cameraOffset[0] + cameraOffset[2] * cameraOffset[2];
+            float distFinish = cameraOffset[0] * cameraOffset[0] + cameraOffset[2] * cameraOffset[2];
             distHold = sqrt(distHold);
             distFinish = sqrt(distFinish);
             distFinish /= distHold;
@@ -227,8 +227,8 @@ void rotateShape(void *gameInfoStruct, void *target) {
         }
         if (currentGame->getKeystate()[SDL_SCANCODE_S] || controllerLeftStateY < -JOYSTICK_DEAD_ZONE) {
             angles[1] = -1.0f * angle - 90.0f;
-            double xSpeed = sin((angles[1]) * (PI / 180)) * speed;
-            double ySpeed = cos((angles[1]) * (PI / 180)) * speed;
+            float xSpeed = sin((angles[1]) * (PI / 180)) * speed;
+            float ySpeed = cos((angles[1]) * (PI / 180)) * speed;
 
             if (controllerLeftStateY < -JOYSTICK_DEAD_ZONE) {
                 multiplier = (static_cast<float>(controllerLeftStateY)) / INT16_MAX;
@@ -238,8 +238,8 @@ void rotateShape(void *gameInfoStruct, void *target) {
         }
         if (currentGame->getKeystate()[SDL_SCANCODE_W] || controllerLeftStateY > JOYSTICK_DEAD_ZONE) {
             angles[1] = -1.0f * angle + 90.0f;
-            double xSpeed = sin(angles[1] * (PI / 180)) * speed;
-            double ySpeed = cos(angles[1] * (PI / 180)) * speed;
+            float xSpeed = sin(angles[1] * (PI / 180)) * speed;
+            float ySpeed = cos(angles[1] * (PI / 180)) * speed;
             if (controllerLeftStateY > JOYSTICK_DEAD_ZONE) {
                 multiplier = (static_cast<float>(controllerLeftStateY * -1)) / INT16_MAX;
             }
@@ -259,8 +259,8 @@ void rotateShape(void *gameInfoStruct, void *target) {
         }
         if (currentGame->getKeystate()[SDL_SCANCODE_A] || controllerLeftStateX > JOYSTICK_DEAD_ZONE) {
             angles[1] = -1.0f * angle + 180.0f;
-            double xSpeed = sin((angles[1]) * (PI / 180)) * speed;
-            double ySpeed = cos((angles[1]) * (PI / 180)) * speed;
+            float xSpeed = sin((angles[1]) * (PI / 180)) * speed;
+            float ySpeed = cos((angles[1]) * (PI / 180)) * speed;
 
             if (controllerLeftStateX > JOYSTICK_DEAD_ZONE) {
                 multiplier = (static_cast<float>(controllerLeftStateX * -1)) / INT16_MAX;
@@ -270,8 +270,8 @@ void rotateShape(void *gameInfoStruct, void *target) {
         }
         if (currentGame->getKeystate()[SDL_SCANCODE_D] || controllerLeftStateX < -JOYSTICK_DEAD_ZONE) {
             angles[1] = -1.0f * angle;
-            double xSpeed = sin((angles[1]) * (PI / 180)) * speed;
-            double ySpeed = cos((angles[1]) * (PI / 180)) * speed;
+            float xSpeed = sin((angles[1]) * (PI / 180)) * speed;
+            float ySpeed = cos((angles[1]) * (PI / 180)) * speed;
 
             if (controllerLeftStateX < -JOYSTICK_DEAD_ZONE) {
                 multiplier = static_cast<float>(controllerLeftStateX) / INT16_MAX;
@@ -336,25 +336,25 @@ void rotateShape(void *gameInfoStruct, void *target) {
 } //NOLINT - refactor required
 
 /*
- (vector<double>) cameraDistance takes a 3D vector containing the offset of the
+ (vector<float>) cameraDistance takes a 3D vector containing the offset of the
  camera from the object and calculates the distance between the two with
  respect to the y-z plane, and x-y plane. Returns a 2D vector where the first
  element is the y-z distance, and the second element is the x-y distance.
 */
-vector<double> cameraDistance(vec3 offset) {
-    vector<double> distance(2);
+vector<float> cameraDistance(vec3 offset) {
+    vector<float> distance(2);
     distance[0] = offset[1] * offset[1] + offset[2] * offset[2];
     distance[1] = offset[1] * offset[1] + offset[0] * offset[0];
     return distance;
 }
 
-double convertNegToDeg(double degree) {
+float convertNegToDeg(float degree) {
     return degree >= 0.0f ? degree : degree + 360.0f;
 }
 
 // Calculates the angle between two points in degrees
-double angleOfPoint(vec3 p1, vec3 p2) {
+float angleOfPoint(vec3 p1, vec3 p2) {
     auto diffPoint = p2 - p1;
-    double angle = atan2(diffPoint[2], diffPoint[0]) * (180.0f / PI);
+    float angle = atan2(diffPoint[2], diffPoint[0]) * (180.0f / PI);
     return convertNegToDeg(angle);
 }
